asiotest: moving crypto bench tests (by Rob/Mikurys) to hpp
